backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asarPy==1.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
colorama==0.4.6
colorlog==6.9.0
cryptography==44.0.0
Cython==3.0.11
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.0
mammoth==1.9.0
Markdown==3.5.2
markdownify==0.14.1
markitdown==0.0.1a3
MarkupSafe==3.0.2
meson==1.5.2
multidict==6.1.0
numpy==1.26.4
openai==1.61.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
pdfminer.six==20240706
pillow==11.1.0
propcache==0.2.1
puremagic==1.28
pycairo==1.27.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
# PyGObject==3.50.0
PyMuPDF==1.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
semantic-router==0.0.72
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.1
starlette==0.45.3
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
XlsxWriter==3.2.2
yarl==1.18.3
youtube-transcript-api==0.6.3
